fix Unit.getUnitsInWeaponRange crashing due to bwapi bug
bwapi's implementation of getUnitsInWeaponRange does not check for a
null filter function, so it crashes if one is not passed in. we were
never passing one in here. to work around this for now, we'll just pass
in a dummy filter function that always returns true (so returns all
units in range always).